created the landing page, the instruction page. i also  gave them styles and responsivness
