+++ src/pages/home.css
@@ -1,0 +1,26 @@
+
+.home-container {
+    text-align: center;
+    padding: 20px;
+    /* border: 2px solid yellow; */
+    margin-top: 6%;
+  }
+  
+  .cta-buttons {
+    margin-top: 20px;
+  }
+  
+  .cta-button {
+    display: inline-block;
+    padding: 10px 20px;
+    background-color: #007bff;
+    color: white;
+    text-decoration: none;
+    border-radius: 5px;
+    font-weight: bold;
+  }
+  
+  .cta-button:hover {
+    background-color: #0056b3;
+  }
+
