+++ src/pages/instructionPage.css
@@ -1,0 +1,58 @@
+.instructions-container {
+    text-align: center;
+    padding: 20px;
+  }
+  
+  .instructions-content {
+    margin-top: 20px;
+    text-align: left;
+  }
+  
+  .instructions-content h3 {
+    margin-bottom: 10px;
+    margin-left: 26%;
+  }
+  
+  .instructions-content ol {
+    padding: 20px;
+    /* border: 2px solid green; */
+    margin-left: 28%;
+
+  }
+  .instructions-content li {
+    margin-bottom: 10px;
+  }
+  
+  .auth-info {
+    margin-top: 30px;
+    font-size: 14px;
+  }
+  
+  .auth-link {
+    color: #007bff;
+    text-decoration: none;
+  }
+  
+  .cta-buttons {
+    margin-top: 30px;
+  }
+  
+  .cta-button {
+    display: inline-block;
+    padding: 10px 20px;
+    background-color: #007bff;
+    color: white;
+    text-decoration: none;
+    border-radius: 5px;
+    font-weight: bold;
+    margin-right: 10px;
+  }
+  
+  .cta-button:last-child {
+    margin-right: 0;
+  }
+  
+  .cta-button:hover {
+    background-color: #0056b3;
+  }
+
